requirements.txt
transformers
tokenizer
datasets
torch
numpy
matplotlib
scipy
json
wandb